Update README.md
This project involves building and evaluating a random forest regression model to classify data from the big sales
dataset.The data is then split into training and testing sets. A random forest regression model is trained on the training data
and used to predict the class labels for the test data. The model’s performance is evaluated.